+++ playgrounds/nuxt/nuxt.config.ts
@@ -5,8 +5,7 @@ const extraAllowedHosts = (process?.env.NUXT_ALLOWED_HOSTS?.split(',').map((s: s
 export default defineNuxtConfig({
   modules: [
     '@bitrix24/b24ui-nuxt',
-    '@bitrix24/b24icons-nuxt',
-    '@nuxtjs/mdc'
+    '@bitrix24/b24icons-nuxt'
   ],
 
   devtools: {